Working Inventory Receipts

+++ internal/database/models/customer.go
@@ -10,8 +10,8 @@ type Customer struct {
 	CustomerID      float64        `bun:"customer_id,pk"`
 	CompanyID       string         `bun:"company_id,pk"`
 	CustomerName    sql.NullString `bun:"customer_name"`
-	CreditLimit     float32        `bun:"credit_limit"`
-	CreditLimitUsed float32        `bun:"credit_limit_used"`
+	CreditLimit     float64        `bun:"credit_limit"`
+	CreditLimitUsed float64        `bun:"credit_limit_used"`
 
 	Orders []*Order `bun:"rel:has-many,join:customer_id=customer_id,join_on:company_id=company_id"`
 }
